+++ scripts/package.js
@@ -32,9 +32,11 @@ function deflate(fname) {
 
 const outputPath = `build/${/^Windows/.test(os.type()) ? 'lumo.exe' : 'lumo'}`;
 const resources = getDirContents('target')
-      .filter(fname => fname.endsWith('.json') ||
-              /clojurescript-version/.test(fname) ||
-              /\$macros\.js$/.test(fname));
+      .filter(fname => !fname.endsWith('main.js') && !fname.endsWith('bundle.js')
+                       && !fname.endsWith('bundle.min.js'));
+      // .filter(fname => fname.endsWith('.json')
+      //         /clojurescript-version/.test(fname) ||
+      //         /\$macros\.js$/.test(fname));
 
 const promises = [];